Bug fix---make sure we handle text files appropriately.
For text files, each line ends with a CRLF, and a final line of just "." is
sent.  If the file isn't text, we send it as is, without the file ".".
RFC-1436 is a bit silent on this matter, and I don't think any clients still
expect an ending ".", but hey, let's try to follow the spec as much as
possible.