Major refactor of code.
I decided to make the handler selectors Lua patterns, like in my Gemini
server.  It allows more flexibility in selecting the appropriate handler.
I'm not entirely happy with this (as I don't really like Lua patterns and
prefer LPEG for parsing needs) but doing this makes it easier for anyone
else to use this.