Return a useable error.
Okay, I'm trying an experiement.  The only thing RFC-1436 says about error
reporting is in the appendix, and in its entirety:

3   Signals an error condition.

That's it!  Nothing more!  So what I'm doing isn't out of spec.  So here's
what I'm doing.  For an error, I'm returning just a fixed set of display
text:

	Not found
	Permanent redirect
	Redirect

The selector, host and port comprise the rest.  For the "Not found" text,
the selector is the selector being requested that is in error.  For both the
"Permanent redirect" and "Redirect" the selector, host and port number are
the new location of the resource in question, of the same type.

Hopefully, this will catch on.