Bug fix---http module.
I handled the reading of the rest of the request via the main loop, but that
code was actually never run due to a module change.  I could have fixed
that, but that was cutting across concerns.  So I made a "private" field in
the request to hold the IO object, so I could move that code to where it's
more of a conern.

This also makes other types of modules easier to implement.