Simplify how selectors are matched.
I decided to remove the Lua regexp patterns I was using.  Now, the selector
field in the handlers specifies the leading text of the selector.  The rest
of the text past this is then passed to the handler.  I think it makes it
much easier to configure, and there's no weird escapes I never remember to
include.