Rework CGI support
I'm rethinking the CGI support.  My thoughts always assuming CGI scripts for
files with the execute bit set are changing.  Now, you need to specify the
cgi field in the handler definition block to true to actually activate CGI
support.  I'm thinking of maying CGI support a sub-module of the filesystem
instead of the pseudo-handler it is now.